+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+### v0.72.1
+- Load artboard image immediately if it intersects the browser viewport.
+
 ### v0.72.0
-- Use IntersectionObserver api in the resizer script, so artboard images are only loaded when they are visible in the browser viewport.
+- Use IntersectionObserver api in the resizer function, so artboard images are only loaded when they are visible in the browser viewport.
 
 ### v0.71.2
 - Support converting multi-segment lines to divs inside :div layers. (Segments must be horizontal or vertical).